refactored test structuring wrt DockerClientPullTest
