+++ client/src/test/de/uniulm/omi/cloudiator/lance/client/ContainersDeployTest.java
@@ -1,0 +1,24 @@
+package de.uniulm.omi.cloudiator.lance.client;
+
+import static de.uniulm.omi.cloudiator.lance.lca.container.ContainerStatus.*;
+import static java.lang.Thread.sleep;
+
+import de.uniulm.omi.cloudiator.domain.OperatingSystemArchitecture;
+import de.uniulm.omi.cloudiator.domain.OperatingSystemFamily;
+import de.uniulm.omi.cloudiator.domain.OperatingSystemImpl;
+import de.uniulm.omi.cloudiator.domain.OperatingSystemVersions;
+import de.uniulm.omi.cloudiator.domain.OperatingSystem;
+
+import de.uniulm.omi.cloudiator.lance.application.component.*;
+import de.uniulm.omi.cloudiator.lance.client.TestUtils.InportInfo;
+import de.uniulm.omi.cloudiator.lance.lca.DeploymentException;
+import java.util.List;
+import org.junit.Test;
+
+import de.uniulm.omi.cloudiator.lance.lifecycle.LifecycleStore;
+import de.uniulm.omi.cloudiator.lance.application.DeploymentContext;
+import de.uniulm.omi.cloudiator.lance.lca.container.ContainerType;
+import java.util.concurrent.Callable;
+
+public class ContainersDeployTest {
+}
